use GitHub Actions for CI/CD

+++ Rakefile
@@ -128,5 +128,5 @@ RDoc::Task.new do |rdoc|
 end
 
 task :clean do
-  sh "rm -f tmp/*.output tmp/*.log tmp/master-dir/* tmp/slave-dir/* tmp/*lock tmp/*pid test.log"
+  sh "rm -f tmp/*.output tmp/*.log tmp/master-dir/* tmp/slave-dir/* tmp/*lock tmp/*pid test.log" unless ENV['GITHUB_ACTIONS']
 end
